add back in the auth login url for ASF
This is the URL that you get redirected to due to a 302 response.

You can see the flow with

$ curl -v -L "https://s1qc.asf.alaska.edu/aux_poeorb/S1A_OPER_AUX_POEORB_OPOD_20231102T080652_V20231012T225942_20231014T005942.EOF"

it was added previously, but was failing in my testing during the big
switchover.

It is possible we don't need this at all if that have an earthdata
token, but i could not get that to work either. Will check with ASF
people about that.