Check listing of EOF files on ASF once only
EOF files on ASF server are listed once only in tests. Ideally this should have
been mocked, but interrogation is still kept.

Other tests will rely on a snapshot of precise orbit files that is cached.
A test dedicated to testing the use of the cached baseline has been defined.